Introduce conditionally-compiled metrics.
This change introduces metrics which are disabled at compile time by default,
but can be enabled using a specified go tag. To demonstrate this, it replaces
the regular KVM-specific metrics to these new ones.

The KVM platform metrics can be enabled with the "condmetric_kvm_profile" tag
when building runsc.

The motivation behind this is twofold:
- We have observed that using metrics in hot-paths can be quite expensive
  (i.e. syscall counter, timer metrics on KVM platform machine.Get()).
- There are classes of metrics that we generally care about using only on
  specific occasions, and don't need to have on by default.

PiperOrigin-RevId: 464151760